// legacy/Reducers/classifications/types.ts
export type IClassificationProperty = {
  Id: number;
  Name: string;
  Description: string;
};

export type IClassification = {
  Id: number;
  Name: string;
  Description: string;
  IsAutomaticTranslate: boolean;
  Version?: string;
  ColorCode?: string;
  IsBimAndCo: boolean;
  IsPrivate: boolean;
  IsMandatory: boolean;
  IsEnabled?: boolean;
  Status?: Status;
  Statistics: {
    ObjectsAssignedPercentage: number;
  };
  CreatedAt: Date;
  UpdatedAt: Date;
  NameProperty: IClassificationProperty;
  CodeProperty: IClassificationProperty;
  Translations?: ITranlation[];
  LanguageCode?: LanguageCode;
};

export type ITranlation = {
  Id?: number;
  LanguageCode?: LanguageCode;
  Name?: string;
  Description?: string;
};

export type IClassificationNode = {
  Id: number;
  Code: string;
  Name: string;
  Description: string;
  ParentId: number;
  IfcExportAsId: null;
  IfcExportTypeId: null;
  CaoCategory: null;
  Children: IClassificationNode[];
  selected?: boolean;
  hasResults?: boolean;
};

export enum Status {
  Public = 'public',
  Private = 'private',
}